✨ Feat: 게시글 수정, 삭제 관련 API 구현 및 테스트 (#48)

+++ src/main/java/HM/Hanbat_Market/service/article/ArticleService.java
@@ -1,20 +1,21 @@
 package HM.Hanbat_Market.service.article;
 
 import HM.Hanbat_Market.api.article.FileStore;
-import HM.Hanbat_Market.api.article.dto.ArticleCreateRequestDto;
-import HM.Hanbat_Market.api.article.dto.ArticleCreateResponseDto;
-import HM.Hanbat_Market.api.article.dto.ArticleDetailResponseDto;
+import HM.Hanbat_Market.api.article.dto.*;
 import HM.Hanbat_Market.api.dto.HomeArticlesDto;
 import HM.Hanbat_Market.domain.entity.Article;
 import HM.Hanbat_Market.domain.entity.ImageFile;
 import HM.Hanbat_Market.domain.entity.Item;
 import HM.Hanbat_Market.domain.entity.Member;
+import HM.Hanbat_Market.exception.ForbiddenException;
+import HM.Hanbat_Market.exception.NotFoundException;
+import HM.Hanbat_Market.exception.member.UnAuthorizedException;
 import HM.Hanbat_Market.repository.article.ArticleRepository;
 import HM.Hanbat_Market.repository.article.dto.ArticleCreateDto;
 import HM.Hanbat_Market.repository.article.dto.ArticleSearchDto;
-import HM.Hanbat_Market.repository.article.dto.ArticleUpdateDto;
 import HM.Hanbat_Market.repository.item.ImageFileRepository;
 import HM.Hanbat_Market.repository.item.dto.ItemCreateDto;
+import HM.Hanbat_Market.repository.item.dto.ItemUpdateDto;
 import HM.Hanbat_Market.repository.member.MemberRepository;
 import HM.Hanbat_Market.service.preemption.PreemptionItemService;
 import lombok.RequiredArgsConstructor;
@@ -41,6 +42,7 @@ public class ArticleService {
     private final ImageFileRepository imageFileRepository;
     private final PreemptionItemService preemptionItemService;
     private final FileStore fileStore = new FileStore();
+    private final String FILE_URL = "https://7d04-39-119-25-167.ngrok-free.app/api/images/";
 
     //영속성 전이로 ItemRepository에 따로 persist하지 않아도 됨
     @Transactional
@@ -57,7 +59,7 @@ public class ArticleService {
     }
 
     @Transactional
-    public ArticleCreateResponseDto createArticleToDto(Long memberId, ArticleCreateRequestDto form, BindingResult result) throws IOException {
+    public ArticleCreateResponseDto createArticleToDto(Long memberId, ArticleCreateRequestDto form) throws IOException {
 
         Member newMember = memberRepository.findById(memberId).get();
 
@@ -76,7 +78,7 @@ public class ArticleService {
 
         // 파일 유효성 검사 실패 시 에러 추가
         if (!multipartFiles.isEmpty()) {
-            List<ImageFile> imageFiles = fileStore.storeFiles(multipartFiles, result);
+            List<ImageFile> imageFiles = fileStore.storeFiles(multipartFiles);
             articleCreateDto.setImageFiles(imageFiles);
         }
 
@@ -90,7 +92,7 @@ public class ArticleService {
                     .forEach(imageFile -> ImageFile
                             .createImageFile(article, imageFile));
         } else if (articleCreateDto.getImageFiles().isEmpty()) {
-            ImageFile imageFile = new ImageFile("default_image", "default_image");
+            ImageFile imageFile = new ImageFile("default_image.png", "default_image.png");
             ImageFile.createImageFile(article, imageFile);
         }
 
@@ -99,9 +101,9 @@ public class ArticleService {
                 article.getTitle(),
                 article.getItem().
                         getItemName(),
-                article.getImageFiles()
+                getFullPath(article.getImageFiles()
                         .get(0)
-                        .getStoreFileName());
+                        .getStoreFileName()));
     }
 
     @Transactional
@@ -126,7 +128,7 @@ public class ArticleService {
                 item.getItemName(),
                 item.getPrice(),
                 member.getNickname(),
-                fileName,
+                getFullPath(fileName),
                 findArticle.getCreatedAt(),
                 preemptionItemSize);
     }
@@ -154,7 +156,7 @@ public class ArticleService {
                             a.getItem().getItemName(),
                             a.getItem().getPrice(),
                             a.getMember().getNickname(),
-                            storeFileName,
+                            getFullPath(storeFileName),
                             a.getCreatedAt()
                     );
                 })
@@ -172,21 +174,90 @@ public class ArticleService {
     }
 
     public Article findArticle(Long articleId) {
+        return articleRepository.findById(articleId)
+                .orElseThrow(() -> new NotFoundException());
+    }
+
+    public Article findUpdateArticle(Long articleId, Member loginMember) {
+
+        Article article = articleRepository.findById(articleId).get();
+
+        if (confirmOwner(article, loginMember)) {
+            throw new ForbiddenException();
+        }
+
         return articleRepository.findById(articleId).get();
     }
 
     @Transactional
-    public void deleteArticle(Long articleId) {
+    public void deleteArticle(Long articleId, Member loginMember) {
+
         Article article = articleRepository.findById(articleId).get();
+
+        if (confirmOwner(article, loginMember)) {
+            throw new ForbiddenException();
+        }
+
         article.delete();
     }
 
     @Transactional
-    public Long updateArticle(Long articleId, ArticleUpdateDto articleUpdateDto) {
+    public ArticleUpdateResponseDto updateArticleToDto(Long articleId,
+                                                       ArticleUpdateRequestDto articleUpdateRequestDto,
+                                                       Member loginMember) throws IOException {
+
+        ArticleUpdateDto articleUpdateDto = new ArticleUpdateDto();
+        ItemUpdateDto itemUpdateDto = new ItemUpdateDto();
+        itemUpdateDto.setItemName(articleUpdateRequestDto.getItemName());
+        itemUpdateDto.setPrice(articleUpdateRequestDto.getPrice());
+        articleUpdateDto.setTitle(articleUpdateRequestDto.getTitle());
+        articleUpdateDto.setDescription(articleUpdateRequestDto.getDescription());
+        articleUpdateDto.setTradingPlace(articleUpdateRequestDto.getTradingPlace());
+
         Article article = articleRepository.findById(articleId).get();
-        article.update(articleUpdateDto);
-        return articleId;
+
+        if (confirmOwner(article, loginMember)) {
+            throw new ForbiddenException();
+        }
+
+        List<MultipartFile> multipartFiles = new ArrayList<>();
+        if (articleUpdateRequestDto.getImageFile1() != null) {
+            multipartFiles.add(articleUpdateRequestDto.getImageFile1());
+        }
+
+        // 파일 유효성 검사 실패 시 예외 발생
+        if (!multipartFiles.isEmpty()) {
+            List<ImageFile> imageFiles = fileStore.storeFiles(multipartFiles);
+            articleUpdateDto.setImageFiles(imageFiles);
+        }
+
+        if (!articleUpdateDto.getImageFiles().isEmpty()) {
+            articleUpdateDto.getImageFiles().stream()
+                    .forEach(imageFile -> ImageFile
+                            .createImageFile(article, imageFile));
+        } else if (articleUpdateDto.getImageFiles().isEmpty()) {
+            ImageFile imageFile = new ImageFile("default_image", "default_image");
+            ImageFile.createImageFile(article, imageFile);
+        }
+
+        article.update(articleUpdateDto, itemUpdateDto);
+
+        return new ArticleUpdateResponseDto(article.getTitle(),
+                article.getDescription(),
+                article.getTradingPlace(),
+                article.getItem().getItemName(),
+                article.getItem().getPrice(),
+                getFullPath(article.getImageFiles().get(0).getStoreFileName()));
     }
 
+    public boolean confirmOwner(Article article, Member member) {
+        if (article.getMember().getId() == member.getId()) {
+            return false;
+        }
+        return true;
+    }
 
+    private String getFullPath(String filename) {
+        return FILE_URL + filename;
+    }
 }
